LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-font-relative-lengths-001-expected.txt


FAIL font-relative lspace refers to the core operator assert_approx_equals: baseSizePx * fontScalePercent * lspaceEm expected 100 +/- 1 but got 50
FAIL font-relative rspace refers to the core operator assert_approx_equals: baseSizePx * fontScalePercent * rspaceEm expected 400 +/- 1 but got 200
FAIL font-relative minsize refers to the core operator assert_approx_equals: baseSizePx * fontScalePercent * minsizeEm expected 400 +/- 1 but got 0
FAIL font-relative maxsize refers to the core operator assert_approx_equals: baseSizePx * fontScalePercent * maxsizeEm expected 200 +/- 1 but got 0
теп

теп
